+++ src/scippnexus/v2/__init__.py
@@ -11,9 +11,6 @@ except importlib.metadata.PackageNotFoundError:
     __version__ = "0.0.0"
 
 from .. import typing
-#from .nxdata import NXdataStrategy
-#from .nxdetector import NXdetectorStrategy
-#from .nxlog import NXlogStrategy
 from .base import (
     Field,
     Group,
@@ -22,8 +19,7 @@ from .base import (
     base_definitions,
     create_class,
     create_field,
-    group_events_by_detector_number,
 )
-#from .definition import ApplicationDefinition, make_definition
 from .file import File
 from .nexus_classes import *
+from .nxdata import group_events_by_detector_number
